refactoring the tabs layout to allow for deep nesting of tabs which will allow us to separate the tabs by concern as opposed to housing them in the same place
