+++ src/features/tabs/tabContent/friendTabs/ProfileFriendMutalsTab.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Container from '../../../../shared/components/container/Container';
+
+const ProfileFriendMutalsTab: React.FC = () => {
+
+     return (
+        <Container>
+            Profile Mutual Friends Tab
+        </Container>
+     );
+};
+export default ProfileFriendMutalsTab;
