+++ src/features/tabs/tabContent/friendTabs/ProfileFriendRequestTab.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Container from '../../../../shared/components/container/Container';
+
+const ProfileFriendRequestTab: React.FC = () => {
+
+     return (
+        <Container>
+            Profile Friends Requests Tab
+        </Container>
+     );
+};
+export default ProfileFriendRequestTab;
