+++ src/features/tabs/tabContent/friendTabs/ProfileFriendsTab.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import Container from '../../../../shared/components/container/Container';
+
+const ProfileFriendsTab: React.FC = () => {
+
+  return (
+    <Container>
+      test
+    </Container>
+  );
+};
+
+export default ProfileFriendsTab;
