+++ src/features/tabs/tabContent/mainTabs/ProfileMainTab.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import Container from '../../../../shared/components/container/Container';
+
+const ProfileMainTab: React.FC = () => {
+
+     return (
+         <Container>
+            Profile Main Tab
+         </Container>
+     );
+};
+export default ProfileMainTab;
