+++ src/features/tabs/tabContent/profileTabs/MyProfileAboutTab.tsx
@@ -1,22 +1,22 @@
 import React, { useEffect, useRef, useState } from "react";
 import type { DragEvent } from "react";
-import Container from "../../../shared/components/container/Container";
-import type { ProfileTab } from "../tabTypes";
-import { useAppDispatch, useAppSelector } from "../../../app/hooks";
-import Text from "../../../shared/components/text/Text";
-import Image from "../../../shared/components/image/Image";
-import Input from "../../../shared/components/input/Input";
-import { openAlert } from "../../alert/alertSlice";
-import { setLoading, setNotLoading } from "../../../app/globalSlices/loading/loadingSlice";
-import { setAuthUser } from "../../auth/authUserSlice";
-import type { AuthUser } from "../../auth/authUserTypes";
-import { updateDataInCollection } from "../../../services/database/updateData";
-import { sendEmailChangeVerification, updateUserPassword } from "../../../services/auth/updateAuthProfile";
-import Button from "../../../shared/components/button/Button";
-import Loader from "../../../shared/components/loader/Loader";
-import { uploadProfileImage } from "../../../services/storage/uploadImage";
-import { deleteImageFromStorage } from "../../../services/storage/deleteImage";
-import Icon from "../../../shared/components/icon/Icon";
+import Container from "../../../../shared/components/container/Container";
+import type { ProfileTab } from "../../tabTypes";
+import { useAppDispatch, useAppSelector } from "../../../../app/hooks";
+import Text from "../../../../shared/components/text/Text";
+import Image from "../../../../shared/components/image/Image";
+import Input from "../../../../shared/components/input/Input";
+import { openAlert } from "../../../alert/alertSlice";
+import { setLoading, setNotLoading } from "../../../../app/globalSlices/loading/loadingSlice";
+import { setAuthUser } from "../../../auth/authUserSlice";
+import type { AuthUser } from "../../../auth/authUserTypes";
+import { updateDataInCollection } from "../../../../services/database/updateData";
+import { sendEmailChangeVerification, updateUserPassword } from "../../../../services/auth/updateAuthProfile";
+import Button from "../../../../shared/components/button/Button";
+import Loader from "../../../../shared/components/loader/Loader";
+import { uploadProfileImage } from "../../../../services/storage/uploadImage";
+import { deleteImageFromStorage } from "../../../../services/storage/deleteImage";
+import Icon from "../../../../shared/components/icon/Icon";
 
 const EMAIL_REGEX = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
 const PHONE_REGEX = /^\d{3}-\d{3}-\d{4}$/;
